# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This page is the reference of the Bacalhau REST API. Project docs are available at https://docs.bacalhau.org/. Find more information about Bacalhau at https://github.com/filecoin-project/bacalhau.",
        "title": "Bacalhau API",
        "contact": {
            "name": "Bacalhau Team",
            "url": "https://github.com/filecoin-project/bacalhau",
            "email": "team@bacalhau.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://github.com/filecoin-project/bacalhau/blob/main/LICENSE"
        }
    },
    "host": "bootstrap.production.bacalhau.org:1234",
    "basePath": "/",
    "paths": {
        "/debug": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Returns debug information on what the current node is doing.",
                "operationId": "apiServer/debug",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/node_info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utils"
                ],
                "summary": "Returns the info of the node.",
                "operationId": "nodeInfo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.NodeInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/requester/debug": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Returns debug information on what the current node is doing.",
                "operationId": "pkg/requester/publicapi/debug",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/requester/events": {
            "post": {
                "description": "Events (e.g. Created, Bid, BidAccepted, ..., ResultsAccepted, ResultsPublished) are useful to track the progress of a job.\n",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Returns the events related to the job-id passed in the body payload. Useful for troubleshooting.",
                "operationId": "pkg/requester/publicapi/events",
                "parameters": [
                    {
                        "description": "Request must specify a `client_id`. To retrieve your `client_id`, you can do the following: (1) submit a dummy job to Bacalhau (or use one you created before), (2) run `bacalhau describe \u003cjob-id\u003e` and fetch the `ClientID` field.",
                        "name": "eventsRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/publicapi.eventsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/publicapi.eventsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/requester/list": {
            "post": {
                "description": "Returns the first (sorted) #`max_jobs` jobs that belong to the `client_id` passed in the body payload (by default).\nIf `return_all` is set to true, it returns all jobs on the Bacalhau network.\n\nIf `id` is set, it returns only the job with that ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Simply lists jobs.",
                "operationId": "pkg/requester/publicapi/list",
                "parameters": [
                    {
                        "description": "Set `return_all` to `true` to return all jobs on the network (may degrade performance, use with care!).",
                        "name": "listRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/publicapi.listRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/publicapi.listResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/requester/local_events": {
            "post": {
                "description": "Local events (e.g. Selected, BidAccepted, Verified) are useful to track the progress of a job.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Returns the node's local events related to the job-id passed in the body payload. Useful for troubleshooting.",
                "operationId": "pkg/requester/publicapi/localEvents",
                "parameters": [
                    {
                        "description": " ",
                        "name": "localEventsRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/publicapi.localEventsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/publicapi.localEventsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/requester/results": {
            "post": {
                "description": "Example response:\n\n```json\n{\n  \"results\": [\n    {\n      \"NodeID\": \"QmdZQ7ZbhnvWY1J12XYKGHApJ6aufKyLNSvf8jZBrBaAVL\",\n      \"Data\": {\n        \"StorageSource\": \"IPFS\",\n        \"Name\": \"job-9304c616-291f-41ad-b862-54e133c0149e-shard-0-host-QmdZQ7ZbhnvWY1J12XYKGHApJ6aufKyLNSvf8jZBrBaAVL\",\n        \"CID\": \"QmTVmC7JBD2ES2qGPqBNVWnX1KeEPNrPGb7rJ8cpFgtefe\"\n      }\n    }\n  ]\n}\n```",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Returns the results of the job-id specified in the body payload.",
                "operationId": "pkg/requester/publicapi/results",
                "parameters": [
                    {
                        "description": " ",
                        "name": "stateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/publicapi.stateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/publicapi.resultsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ComputeNodeInfo": {
            "type": "object",
            "properties": {
                "AvailableCapacity": {
                    "$ref": "#/definitions/model.ResourceUsageData"
                },
                "EnqueuedExecutions": {
                    "type": "integer"
                },
                "ExecutionEngines": {
                    "type": "array",
                    "items": {
                        "description": "e.g. docker or language",
                        "type": "integer"
                    }
                },
                "MaxCapacity": {
                    "$ref": "#/definitions/model.ResourceUsageData"
                },
                "MaxJobRequirements": {
                    "$ref": "#/definitions/model.ResourceUsageData"
                },
                "RunningExecutions": {
                    "type": "integer"
                }
            }
        },
...
        "model.Job": {
            "type": "object",
            "properties": {
                "APIVersion": {
                    "type": "string",
                    "example": "V1beta1"
                },
                "Metadata": {
                    "$ref": "#/definitions/model.Metadata"
                },
                "Spec": {
                    "description": "The specification of this job.",
                    "$ref": "#/definitions/model.Spec"
                },
                "Status": {
                    "description": "The status of the job: where are the nodes at, what are the events",
                    "$ref": "#/definitions/model.JobStatus"
                }
            }
        },
        "model.JobEvent": {
            "type": "object",
            "properties": {
                "APIVersion": {
                    "description": "APIVersion of the Job",
                    "type": "string",
                    "example": "V1beta1"
                },
                "ClientID": {
                    "description": "optional clientID if this is an externally triggered event (like create job)",
                    "type": "string",
                    "example": "ac13188e93c97a9c2e7cf8e86c7313156a73436036f30da1ececc2ce79f9ea51"
                },
                "Deal": {
                    "description": "this is only defined in \"update_deal\" events",
                    "$ref": "#/definitions/model.Deal"
                },
                "EventName": {
                    "type": "integer"
                },
                "EventTime": {
                    "type": "string",
                    "example": "2022-11-17T13:32:55.756658941Z"
                },
                "ExecutionID": {
                    "description": "compute execution identifier",
                    "type": "string",
                    "example": "9304c616-291f-41ad-b862-54e133c0149e"
                },
                "JobExecutionPlan": {
                    "description": "this is only defined in \"create\" events",
                    "$ref": "#/definitions/model.JobExecutionPlan"
                },
                "JobID": {
                    "type": "string",
                    "example": "9304c616-291f-41ad-b862-54e133c0149e"
                },
                "PublishedResult": {
                    "$ref": "#/definitions/model.StorageSpec"
                },
                "RunOutput": {
                    "description": "RunOutput of the job",
                    "$ref": "#/definitions/model.RunCommandResult"
                },
                "SenderPublicKey": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "ShardIndex": {
                    "description": "what shard is this event for",
                    "type": "integer"
                },
                "SourceNodeID": {
                    "description": "the node that emitted this event",
                    "type": "string",
                    "example": "QmXaXu9N5GNetatsvwnTfQqNtSeKAD6uCmarbh3LMRYAcF"
                },
                "Spec": {
                    "description": "this is only defined in \"create\" events",
                    "$ref": "#/definitions/model.Spec"
                },
                "Status": {
                    "type": "string",
                    "example": "Got results proposal of length: 0"
                },
                "TargetNodeID": {
                    "description": "the node that this event is for\ne.g. \"AcceptJobBid\" was emitted by Requester but it targeting compute node",
                    "type": "string",
                    "example": "QmdZQ7ZbhnvWY1J12XYKGHApJ6aufKyLNSvf8jZBrBaAVL"
                },
                "VerificationProposal": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "VerificationResult": {
                    "$ref": "#/definitions/model.VerificationResult"
                }
            }
        },
        "model.JobExecutionPlan": {
            "type": "object",
            "properties": {
                "ShardsTotal": {
                    "description": "how many shards are there in total for this job\nwe are expecting this number x concurrency total\nJobShardState objects for this job",
                    "type": "integer"
                }
            }
        },
        "model.JobLocalEvent": {
            "type": "object",
            "properties": {
                "EventName": {
                    "type": "integer"
                },
                "JobID": {
                    "type": "string"
                },
                "ShardIndex": {
                    "type": "integer"
                },
                "TargetNodeID": {
                    "type": "string"
                }
            }
        },
...
        "model.NodeInfo": {
            "type": "object",
            "properties": {
                "ComputeNodeInfo": {
                    "$ref": "#/definitions/model.ComputeNodeInfo"
                },
                "Labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "NodeType": {
                    "type": "integer"
                },
                "PeerInfo": {
                    "$ref": "#/definitions/peer.AddrInfo"
                }
            }
        },
        "model.PublishedResult": {
            "type": "object",
            "properties": {
                "Data": {
                    "$ref": "#/definitions/model.StorageSpec"
                },
                "NodeID": {
                    "type": "string"
                },
                "ShardIndex": {
                    "type": "integer"
                }
            }
        },
...
        "model.ResourceUsageData": {
            "type": "object",
            "properties": {
                "CPU": {
                    "description": "cpu units",
                    "type": "number",
                    "example": 9.600000000000001
                },
                "Disk": {
                    "description": "bytes",
                    "type": "integer",
                    "example": 212663867801
                },
                "GPU": {
                    "type": "integer",
                    "example": 1
                },
                "Memory": {
                    "description": "bytes",
                    "type": "integer",
                    "example": 27487790694
                }
            }
        },
...
        "peer.AddrInfo": {
            "type": "object",
            "properties": {
                "addrs": {
                    "type": "array",
                    "items": {}
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "publicapi.eventsRequest": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string",
                    "example": "ac13188e93c97a9c2e7cf8e86c7313156a73436036f30da1ececc2ce79f9ea51"
                },
                "job_id": {
                    "type": "string",
                    "example": "9304c616-291f-41ad-b862-54e133c0149e"
                }
            }
        },
        "publicapi.eventsResponse": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.JobEvent"
                    }
                }
            }
        },
        "publicapi.listRequest": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string",
                    "example": "ac13188e93c97a9c2e7cf8e86c7313156a73436036f30da1ececc2ce79f9ea51"
                },
                "exclude_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['any-tag']"
                    ]
                },
                "id": {
                    "type": "string",
                    "example": "9304c616-291f-41ad-b862-54e133c0149e"
                },
                "include_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['any-tag']"
                    ]
                },
                "max_jobs": {
                    "type": "integer",
                    "example": 10
                },
                "return_all": {
                    "type": "boolean"
                },
                "sort_by": {
                    "type": "string",
                    "example": "created_at"
                },
                "sort_reverse": {
                    "type": "boolean"
                }
            }
        },
        "publicapi.listResponse": {
            "type": "object",
            "properties": {
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Job"
                    }
                }
            }
        },
        "publicapi.localEventsRequest": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                }
            }
        },
        "publicapi.localEventsResponse": {
            "type": "object",
            "properties": {
                "localEvents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.JobLocalEvent"
                    }
                }
            }
        },
        "publicapi.resultsResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PublishedResult"
                    }
                }
            }
        },
        "publicapi.stateRequest": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string",
                    "example": "ac13188e93c97a9c2e7cf8e86c7313156a73436036f30da1ececc2ce79f9ea51"
                },
                "job_id": {
                    "type": "string",
                    "example": "9304c616-291f-41ad-b862-54e133c0149e"
                }
            }
        },
...
    }
}